Remove --no-build flag and clean up project dependencies
Removed the `--no-build` flag from `dotnet pack` commands in `deterministic-build.yml` to ensure the project is built during
packaging. This affects both the initial package creation and verification steps.

In `Codebreaker.Analyzers.csproj`, removed the `<WarningsAsErrors />`  element, allowing warnings to no longer be treated as errors.
Also removed the `<PackageReference>` to `System.Text.Json`, as this was a workaround for the project or NuGet packaging.